Move information extraction to clinlp.ie
