+++ src/clinlp/__init__.py
@@ -1,13 +1,16 @@
 import importlib.util
 
-from .entity import EntityMatcher, Term
-from .language import Clinlp
-from .normalizer import Normalizer
-from .qualifier import ContextAlgorithm
-from .sentencizer import Sentencizer
+# components
+from .ie.entity import EntityMatcher
+from .ie.qualifier import ContextAlgorithm
 
 if importlib.util.find_spec("transformers") is not None:
-    from .qualifier import NegationTransformer
+    from .ie.qualifier import ExperiencerTransformer, NegationTransformer
 
-__all__ = ['EntityMatcher', 'Term', 'Clinlp', 'Normalizer', 'ContextAlgorithm',
-           'Sentencizer', 'NegationTransformer', 'ExperiencerTransformer']
+# base
+from .language import Clinlp
+from .normalizer import Normalizer
+from .sentencizer import Sentencizer
+
+__all__ = ['EntityMatcher', 'ContextAlgorithm', 'ExperiencerTransformer',
+           'NegationTransformer', 'Clinlp', 'Normalizer', 'Sentencizer']
