+++ src/clinlp/ie/__init__.py
@@ -1,0 +1,3 @@
+from .entity import EntityMatcher, Term, create_concept_dict
+
+__all__ = ["EntityMatcher", "Term", "create_concept_dict"]
