+++ tests/regression/ie/qualifier/test_regression_transformer_negation.py
@@ -4,13 +4,12 @@ import pytest
 import spacy
 
 import clinlp  # noqa: F401
-from clinlp.qualifier.qualifier import ATTR_QUALIFIERS_STR
+from clinlp.ie.qualifier.qualifier import ATTR_QUALIFIERS_STR
 
 
 @pytest.fixture()
 def nlp():
     nlp = spacy.blank("clinlp")
-    nlp.add_pipe("clinlp_normalizer")
     nlp.add_pipe("clinlp_sentencizer")
 
     # ruler
@@ -18,12 +17,14 @@ def nlp():
     ruler.add_patterns([{"label": "named_entity", "pattern": "ENTITY"}])
 
     # recognizer
-    _ = nlp.add_pipe("clinlp_context_algorithm", config={"phrase_matcher_attr": "NORM"})
+    _ = nlp.add_pipe(
+        "clinlp_negation_transformer", config={"token_window": 32, "placeholder": "X"}
+    )
 
     return nlp
 
 
-class TestRegressionContextAlgorithm:
+class TestRegressionTransformer:
     def test_qualifier_cases(self, nlp):
         with open("tests/data/qualifier_cases.json", "rb") as file:
             data = json.load(file)
@@ -40,18 +41,17 @@ class TestRegressionContextAlgorithm:
                     assert predicted_ent.start == example_ent["start"]
                     assert predicted_ent.end == example_ent["end"]
                     assert str(predicted_ent) == example_ent["text"]
-                    assert getattr(predicted_ent._, ATTR_QUALIFIERS_STR).issuperset(
-                        set(example_ent["qualifiers"])
+                    assert getattr(predicted_ent._, ATTR_QUALIFIERS_STR).issubset(
+                        example_ent["qualifiers"]
                     )
-
                 except AssertionError:
                     print(
                         f"Incorrect (#{example_ent['ent_id']}): "
-                        f"text={example['text']}, "
-                        f"example_ent={example_ent}, "
+                        f"text="
+                        f"{example['text']}, example_ent={example_ent}, "
                         f"predicted qualifiers="
                         f"{getattr(predicted_ent._, ATTR_QUALIFIERS_STR)}"
                     )
                     incorrect_ents.add(example_ent["ent_id"])
 
-        assert incorrect_ents == set()
+        assert incorrect_ents == {9}
